# server/api/openapi.yaml
openapi: 3.1.0
info:
  title: Argus API
  version: "1.0.0"
  description: >-
    Argus's read API (docs/SPEC.md §4): agent-telemetry sessions, timelines,
    events, tool calls, subagent trees, and rollup-backed analytics, plus the
    ops endpoints and the ingest surface (OTLP + the hooks webhook). This
    file is hand-authored and reviewed like code (§4.4): it is the contract
    handlers must satisfy, not a byproduct of them.

    Every vendor-supplied field (project, model, vendor, decision_source,
    query_source, tool_source, permission_mode, start_type, terminal_type,
    and friends) is typed `string` with no `enum`, on purpose (SPEC §0,
    §4.4): Argus ingests values it has never seen and must never reject or
    mistype them. Only Argus's own four closed taxonomies — `kind`,
    `source`, `correlation`, `status` (SPEC §0) — and Argus-invented request
    parameters (sort keys, dimensions, buckets) are enumerated.
servers:
  - url: /
    description: Argus serves the ops, read, and ingest surfaces from one origin.
...
paths:
...
  /api/v1/sessions/{id}:
    get:
      operationId: getSession
      tags: [sessions]
      summary: Session detail with aggregates
      parameters:
        - $ref: "#/components/parameters/SessionID"
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: The session, plus permission history, top tools, and decision summary.
          headers:
            ETag:
              description: Hash of the underlying max(ts, seq) + filter (SPEC §4.1).
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDetail"
              examples:
                default:
                  value:
                    id: "3f7a3b1e-0000-0000-0000-000000000001"
                    vendor: "claude_code"
                    project: "argus"
                    cwd: "/home/y/Labs/argus"
                    status: "active"
                    start_type: "fresh"
                    started_at: "2026-08-11T09:02:11.412Z"
                    ended_at: null
                    last_event_at: "2026-08-11T09:31:44.900Z"
                    duration_ms: 1773488
                    turn_count: 12
                    event_count: 480
                    tool_call_count: 96
                    tool_reject_count: 3
                    subagent_count: 2
                    error_count: 1
                    tokens: { input: 41233, output: 18944, cache_read: 1204331, cache_creation: 88210 }
                    cost:
                      usd: 4.2711
                      reported_usd: 4.2711
                      estimated_usd: 0
                      estimated_share: 0
                      by_query_source: { "": 3.9011, "sdk": 0.35 }
                      dominant_query_source: ""
                      other_query_source_usd: 0.35
                    models: ["claude-opus-5"]
                    partial: false
                    app_version: "2.1.220"
                    entrypoint: "cli"
                    terminal_type: "wsl-Ubuntu"
                    permission_mode_history:
                      - ts: "2026-08-11T09:05:00.000Z"
                        from: "default"
                        to: "acceptEdits"
                        trigger: "user"
                    top_tools:
                      - { tool_name: "Edit", calls: 40, rejects: 2, p50_ms: 120 }
                    decision_summary:
                      accept: 90
                      reject: 6
                      by_source: { config: 60, hook: 10, user_permanent: 10, user_temporary: 5, user_reject: 4, user_abort: 1 }
                      exact_share: 1.0
                    sources_seen: ["otel_log", "hook"]
                    raw_events_expired: false
                    hook_latency: { p50_ms: 9, p95_ms: 41, by_hook_event: { PostToolUse: 9 } }
                    first_seen_at: "2026-08-11T09:02:10.000Z"
                    user: "yohann"
                    organization_id: "org_1"
        "304":
          description: Not modified — the client's If-None-Match matches the current ETag.
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
components:
  parameters:
    SessionID:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    IfNoneMatch:
      name: If-None-Match
      in: header
      description: ETag from a previous response; a match returns 304 with an empty body.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Malformed or out-of-range request parameter.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            invalidCursor:
              value:
                type: "urn:argus:error:invalid-cursor"
                title: "Bad Request"
                status: 400
                detail: "cursor failed signature check"
                instance: "/api/v1/sessions"
            invalidParameter:
              value:
                type: "urn:argus:error:invalid-parameter"
                title: "Bad Request"
                status: 400
                detail: "from: not a valid RFC 3339 timestamp or relative shorthand: \"garbage\""
                instance: "/api/v1/sessions"
    Unauthorized:
      description: Missing or invalid bearer token (ARGUS_API_TOKEN configured).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            default:
              value:
                type: "urn:argus:error:unauthorized"
                title: "Unauthorized"
                status: 401
                detail: "missing or invalid bearer token"
                instance: "/api/v1/sessions"
    NotFound:
      description: No such resource.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            default:
              value:
                type: "urn:argus:error:not-found"
                title: "Not Found"
                status: 404
                detail: "no such resource"
                instance: "/api/v1/sessions/does-not-exist"
    MethodNotAllowed:
      description: Method not allowed on this resource.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            default:
              value:
                type: "urn:argus:error:method-not-allowed"
                title: "Method Not Allowed"
                status: 405
                detail: "method not allowed on this resource"
                instance: "/api/v1/sessions"
...
  schemas:
...
    Problem:
      type: object
      description: >-
        RFC 9457 problem+json (SPEC §4.1). type is a stable
        urn:argus:error:<slug> from Argus's own error catalog (see
        server/internal/httpapi/problem.go); left as a plain string rather
        than enumerated here so a future URN never needs a spec-only
        migration to become valid.
      required: [type, title, status]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: array
          items:
            type: object
            additionalProperties: true
...
    EventSource:
      type: string
      description: Provenance pipeline (SPEC §1.3) — Argus-closed, one of the four permitted taxonomies.
      enum: [otel_log, otel_metric, hook, sim]
...
    SessionStatus:
      type: string
      description: Argus-computed session lifecycle state (SPEC §1.7) — closed, unlike vendor vocabulary.
      enum: [active, ended, abandoned, unknown]
...
    TokenUsage:
      type: object
      required: [input, output, cache_read, cache_creation]
      properties:
        input: { type: integer, format: int64 }
        output: { type: integer, format: int64 }
        cache_read: { type: integer, format: int64 }
        cache_creation: { type: integer, format: int64 }
    SessionCost:
      type: object
      description: >-
        by_query_source keys are raw agent-supplied query_source values,
        passed through verbatim (SPEC §1.9) — never mapped onto a
        main/subagent semantic.
      required: [usd, reported_usd, estimated_usd, estimated_share, by_query_source, dominant_query_source, other_query_source_usd]
      properties:
        usd: { type: number }
        reported_usd: { type: number }
        estimated_usd: { type: number }
        estimated_share: { type: number }
        by_query_source:
          type: object
          additionalProperties:
            type: number
        dominant_query_source:
          type: string
        other_query_source_usd:
          type: number
    SessionSummary:
      type: object
      required:
        [id, vendor, project, cwd, status, start_type, started_at, ended_at, last_event_at, duration_ms,
          turn_count, event_count, tool_call_count, tool_reject_count, subagent_count, error_count,
          tokens, cost, models, partial, app_version, entrypoint, terminal_type]
      properties:
        id: { type: string }
        vendor:
          type: string
          description: Agent-agnostic core text (claude_code, codex, gemini_cli, unknown); deliberately unconstrained.
        project: { type: string }
        cwd: { type: string }
        status:
          $ref: "#/components/schemas/SessionStatus"
        start_type:
          type: string
          description: Vendor-supplied session start reason; unconstrained (SPEC §0).
        started_at:
          type: [string, "null"]
          format: date-time
          description: NULL until session.start is seen (SPEC §1.7).
        ended_at:
          type: [string, "null"]
          format: date-time
        last_event_at:
          type: string
          format: date-time
        duration_ms:
          type: [integer, "null"]
          format: int64
        turn_count: { type: integer }
        event_count: { type: integer, format: int64 }
        tool_call_count: { type: integer }
        tool_reject_count: { type: integer }
        subagent_count: { type: integer }
        error_count: { type: integer }
        tokens:
          $ref: "#/components/schemas/TokenUsage"
        cost:
          $ref: "#/components/schemas/SessionCost"
        models:
          type: array
          items:
            type: string
        partial:
          type: boolean
          description: True when no session.start was ever seen (stub-on-reference, SPEC §1.7).
        app_version:
          type: string
        entrypoint:
          type: string
        terminal_type:
          type: string
          description: Vendor-supplied, free-form; a live capture found "wsl-Ubuntu" outside any documented enum.
...
    PermissionModeChange:
      type: object
      required: [ts, from, to, trigger]
      properties:
        ts: { type: string, format: date-time }
        from:
          type: string
          description: Vendor-supplied permission_mode value; unconstrained.
        to:
          type: string
          description: Vendor-supplied permission_mode value; unconstrained.
        trigger:
          type: string
    ToolUsageSummary:
      type: object
      required: [tool_name, calls, rejects, p50_ms]
      properties:
        tool_name: { type: string }
        calls: { type: integer }
        rejects: { type: integer }
        p50_ms:
          type: [integer, "null"]
    SessionDecisionSummary:
      type: object
      required: [accept, reject, by_source, exact_share]
      properties:
        accept: { type: integer }
        reject: { type: integer }
        by_source:
          type: object
          description: Keys are raw decision_source values, unconstrained (SPEC §1.9).
          additionalProperties:
            type: integer
        exact_share: { type: number }
    SessionHookLatency:
      type: object
      required: [p50_ms, p95_ms, by_hook_event]
      properties:
        p50_ms: { type: integer, format: int64 }
        p95_ms: { type: integer, format: int64 }
        by_hook_event:
          type: object
          additionalProperties:
            type: integer
            format: int64
    SessionDetail:
      allOf:
        - $ref: "#/components/schemas/SessionSummary"
        - type: object
          required:
            [permission_mode_history, top_tools, decision_summary, sources_seen,
              raw_events_expired, hook_latency, first_seen_at, user, organization_id]
          properties:
            permission_mode_history:
              type: array
              items:
                $ref: "#/components/schemas/PermissionModeChange"
            top_tools:
              type: array
              items:
                $ref: "#/components/schemas/ToolUsageSummary"
            decision_summary:
              $ref: "#/components/schemas/SessionDecisionSummary"
            sources_seen:
              type: array
              items:
                $ref: "#/components/schemas/EventSource"
            raw_events_expired:
              type: boolean
            hook_latency:
              oneOf:
                - $ref: "#/components/schemas/SessionHookLatency"
                - type: "null"
            first_seen_at:
              type: string
              format: date-time
            user:
              type: string
            organization_id:
              type: string
...